Bug fixes - scoping #27
This is to fix many bugs with the real time interpreting feature.
Some of these issues were not obvious. Such as how Rowdy looks for variables in the current scope. Since there is no main to execute and Rowdy executes single lines as simply statements this caused some issues with scope and how/where variables are allocated.